% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CorTable2.R
\name{CorTable2}
\alias{CorTable2}
\title{APA Correlation Table}
\usage{
CorTable2(x, caption = "", notes = "", alpha = 0.05, digits = 2,
  boldSig = TRUE, includeMeanSD = TRUE, includeCaptionNumber = TRUE,
  WidthProportion = 1, use = "pairwise.complete.obs", landscape = FALSE)
}
\arguments{
\item{x}{A data frame with numeric values}
}
\value{
APA correlation table
}
\description{
APA correlation tables are created from data frames using CorTable
}

