% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcolumn_format.R
\name{dcolumn_format}
\alias{dcolumn_format}
\title{Format for dcolumn}
\usage{
dcolumn_format(x, splitter = ".")
}
\arguments{
\item{x}{a character vector}

\item{splitter}{the character on which to align the column}
}
\description{
Width of column to the right and left of decimal
}
