% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pvalueAPA.R
\name{pvalueAPA}
\alias{pvalueAPA}
\title{APA Style P-value}
\usage{
pvalueAPA(p, inline = FALSE, mindigits = 2, maxdigits = 3)
}
\arguments{
\item{p}{a numeric value between 0 and 1}

\item{inline}{If TRUE returns LaTeX equation (e.g., $p=.02$). If false (default), only the number is returned.}

\item{mindigits}{The minimum number digits for rounding.}

\item{maxdigits}{The maximum number of digits for round small numbers.}
}
\value{
APA formatted p-value
}
\description{
P-values are formatted to APA format from a numeric value using pvalueAPA
}

