% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_coef.R
\name{reg_coef}
\alias{reg_coef}
\title{Regression Analysis Summary Table}
\usage{
reg_coef(m, caption = "Regression Analysis Summary", digits = 2,
  level = 0.95, markdown = TRUE, unstandardized = TRUE, se = TRUE,
  ci = TRUE, t.value = TRUE, p.value = TRUE, standardized = TRUE,
  zero.order.cor = FALSE, product.measure = FALSE,
  structure.coef = FALSE, partial.cor = FALSE,
  semipartial.cor = FALSE, semipartial.cor.squared = TRUE)
}
\arguments{
\item{m}{An object of class lm.}

\item{caption}{A character string denoting the title of the regression table.}

\item{digits}{Number of digits after the decimal place. Defaults to 2.}

\item{level}{Set the confidence interval. Defaults to 0.95.}

\item{markdown}{Return a markdown table. Defaults to TRUE.}

\item{unstandardized}{Include unstandardized estimates. Defaults to TRUE.}

\item{se}{Include standard error. Defaults to TRUE.}

\item{ci}{Include confidence interval. Defaults to TRUE.}

\item{t.value}{Include t-values. Defaults to TRUE.}

\item{p.value}{Include p-values. Defaults to TRUE.}

\item{standardized}{Include standardized estimates. Defaults to TRUE.}

\item{zero.order.cor}{Include zero-order correlations. Defaults to FALSE.}

\item{product.measure}{Include the product measure (zero-order correlation * beta)}

\item{structure.coef}{Include structure coefficients. Defaults to FALSE.}

\item{partial.cor}{Include partial correlations. Defaults to FALSE.}

\item{semipartial.cor}{Include semi-partial correlations. Defaults to FALSE.}

\item{semipartial.cor.squared}{Include semi-partial correlations squared. Defaults to TRUE}
}
\description{
Regression Analysis Summary Table
}
