% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnum.R
\name{tabnum}
\alias{tabnum}
\title{Format numbers for tables by replacing the hyphen with a minus sign.}
\usage{
tabnum(x, digits = 2, stripzero = FALSE, addDollars = FALSE)
}
\arguments{
\item{x}{A numeric variable.}

\item{digits}{Number of digits after the decimal place. Default = 2}

\item{stripzero}{Remove the leading zero? Default = FALSE}

\item{addDollars}{Add dollar signs for latex math mode}
}
\description{
Format numbers for tables by replacing the hyphen with a minus sign.
}
