% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termCleaner.R
\name{termCleaner}
\alias{termCleaner}
\title{Interaction term replace}
\usage{
termCleaner(term)
}
\arguments{
\item{term}{A vector of effect names}
}
\description{
Replaces ":" in interaction terms with the LaTeX multiply symbol
}
