% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreePartTable.R
\name{ThreePartTable}
\alias{ThreePartTable}
\title{Three-part tables (Caption-body-note)}
\usage{
ThreePartTable(tablebody, caption = "", note = "", align = NA,
  digits = 2)
}
\arguments{
\item{tablebody}{A matrix or data.frame of the table.}

\item{caption}{A character string denoting the title of the table.}

\item{note}{A character string denoting the note at the foot of the table.}

\item{align}{A character string to align columns: l = left, c = center, r = right, d = decimal e.g., align = "llccrrdd"}

\item{digits}{A numeric value indicating the number of digits to round numeric columns}
}
\description{
Three-part tables (Caption-body-note)
}
